+++ index.d.ts
@@ -4,6 +4,21 @@ type ValueType =
 	Number |
 	Boolean;
 
+// It's unclear how to express something like `type? = Type` in TypeScript.
+// So instead it's defined as `type?: CellType<Type>`.
+// https://gitlab.com/catamphetamine/write-excel-file/-/issues/4#note_715204034
+// https://www.typescriptlang.org/docs/handbook/2/conditional-types.html
+type CellType<Type> =
+	Type extends String
+		? StringConstructor
+		: Type extends Date
+			? DateConstructor
+			: Type extends Number
+				? NumberConstructor
+				: Type extends Boolean
+					? BooleanConstructor
+					: never
+
 type BorderStyle =
 	'hair' |
 	'dotted' |
@@ -46,17 +61,20 @@ interface CellStyle {
 }
 
 interface CellProps<Type> extends CellStyle {
-	type?: Type;
+	// It's unclear how to express something like `type? = Type` in TypeScript.
+	type?: CellType<Type>;
+	// A simpler (loose) variant:
+	// type?: ValueType;
 	format?: string;
 }
 
 interface CellOfType<Type> extends CellProps<Type> {
-	value?: typeof Type;
+	value?: Type;
 }
 
 type Cell = CellOfType<ValueType>
 
-type Row = Cell[];
+export type Row = Cell[];
 
 export type SheetData = Row[];
 
@@ -66,7 +84,7 @@ interface ColumnSchema<Object, Type> extends CellProps<Type> {
 	// Column width (in characters).
 	width?: number;
 	// Cell value getter.
-	value(object: Object): typeof Type | undefined | null;
+	value(object: Object): Type | undefined | null;
 }
 
 export type Schema<Object> = ColumnSchema<Object, ValueType>[];
@@ -78,7 +96,7 @@ type Column = {
 export type Columns = Column[];
 
 export interface CommonOptions {
-  headerStyle?: CellProps;
+  headerStyle?: CellProps<ValueType>;
   fontFamily?: string;
   fontSize?: number;
   dateFormat?: string;
@@ -92,7 +110,7 @@ interface WithSchemaOptions<Object> extends CommonOptions {
 	fileName: string;
 }
 
-function writeXlsxFile<Object>(
+declare function writeXlsxFile<Object>(
 	objects: Object[] | Object[][],
 	options: WithSchemaOptions<Object>
 ) : Promise<void>;
@@ -105,7 +123,7 @@ interface WithoutSchemaOptions extends CommonOptions {
 	fileName: string;
 }
 
-function writeXlsxFile(
+declare function writeXlsxFile(
 	data: SheetData | SheetData[],
 	options: WithoutSchemaOptions
 ) : Promise<void>;